Refactor Transcoder class and update CMake configuration for conditional transcoder selection
Refactor Transcoder class and update CMake configuration for conditional transcoder selection

fix bmf transcoder libraries link method

+++ src/common/include/process_parameter.h
@@ -47,7 +47,7 @@ class ProcessParameter : public QObject {
     void update_Time_Required(double result);
 
   private:
-    double processNumber;
+    int64_t processNumber;
 
     double timeRequired;
 };
